#2 create certificate as k8s secret
earlier certificate file paths have been removed in favor of actual binary data
without file reference. This led to errors with the start-up probe which still
likes file references.

In order to re-gain files k8s-ces-control creates a TLS secret along with the
certificate (which is already added to the registry). This secret should be
mounted as file so the start-up probes can work as usual.