#2 fix certificate update problem
In earlier versions of the deployment, the mounted secret did not reflect
changes when the underlying secret was updated. That in turn led to buggy start-
ups in which either the certificate did not exist in the file system or the
GRPC service did not return successful probes.

This commit now mounts the secret in a way that when updated the changes are
reflected immediately in the file system.

Also the secret will now updated upon every start to reflect changes of the
etcd-stored certificate.