#25 Decouple grpc-request context from RestartDogu context by creating a new context
This was added to prevent a bug when restarting the admin dogu. The context was cancelled as soon as the deployment was scaled to 0 and then the restart was also cancelled at this point so the deployment of admin dogu remained at 0. By using context.Background(), this bug was fixed.