#64 Deactivate node exporter to fix setup in ci
Since we only want to test if an archive is created, it is ok to disable the node exporter. For now.

In the future, it can happen that the creation of the archive fails because the node metrics are not available.
In this case, we either have to exclude the node metrics from the archive or mount / in the k3d container with the k3d cluster create command.
After that, the mount from the node exporter has to be edited to the appropriate mount paths.